+++ src/scalib/postprocessing/__init__.py
@@ -1,1 +1,17 @@
+r"""
+Rank Estimation
+^^^^^^^^^^^^^^^
+
+.. currentmodule:: scalib.postprocessing
+
+.. autosummary::
+   :toctree:
+   :recursive:
+   :nosignatures:
+
+   scalib.postprocessing.rankestimation
+"""
+
+__all__ = ["rankestimation"]
+
 from .rankestimation import rank_nbin, rank_accuracy
