+++ docs/source/scalib.postprocessing.rst
@@ -1,9 +1,0 @@
-Postprocessing
-==============
-
-RankEstimation
---------------
-
-.. automodule:: scalib.postprocessing.rankestimation
-   :members:
-   :undoc-members:
