+++ docs/source/scalib.metrics.rst
@@ -1,13 +1,0 @@
-Metrics
-=======
-
-This module contains several well-know side-channel metrics.
-
-.. toctree::
-   :caption: Available metrics are:
-   :maxdepth: 1
-
-   scalib.metrics.snr.rst 
-   scalib.metrics.ttest.rst
-   
-
