+++ docs/source/scalib.configuration.rst
@@ -1,6 +1,0 @@
-Configuration
-=============
-
-.. automodule:: scalib.config
-   :members:
-   :undoc-members:
