+++ docs/source/copyright.rst
@@ -1,0 +1,4 @@
+Copyright
+=========
+
+.. include:: ../../COPYRIGHT
