+++ docs/source/api_ref.rst
@@ -1,0 +1,17 @@
+API Reference
+=============
+
+.. currentmodule:: scalib
+
+
+.. autosummary::
+   :recursive:
+   :nosignatures:
+   :toctree: _generated
+
+   metrics
+   modeling
+   attacks
+   postprocessing
+   config
+
