+++ docs/source/scalib.modeling.rst
@@ -1,8 +1,0 @@
-Modeling
-========
-
-LDAClassifier
--------------
-
-.. autoclass:: scalib.modeling.LDAClassifier
-   :members:
