+++ docs/source/scalib.metrics.snr.rst
@@ -1,7 +1,0 @@
-Signal-to-Noise Ratio (SNR)
-===========================
-
-.. autoclass:: scalib.metrics.SNR
-   :members:
-
-

[…]
