### 2023-02-16 #5158 @cassiersg
As potential reviewers, the following researchers are experts in side-channel analysis and have published open-source software:

```
nicolaimueller
JannikZeitschner
gs1989
byt3bit
```

In the list of people who agreed to review for JOSS, I would suggest `ceb8`.
