### 2023-02-17 #5158 @Nikoleta-v3
👋🏻 @nicolaimueller @JannikZeitschner @gs1989 @byt3bit would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "SCALib: A Side-Channel Analysis Library". You can find more information at the top of this Github issue 🆙
