### 2023-02-20 d43cb29
New methods on FactorGraph and BPState
* Add ``vars`` and ``factors`` methods on `FactorGraph`.
* Add ``fg`` property on `BPState`.
