### 2023-02-20 d912f13
Handle negative values below MIN_PROBA in belief propagation.
This was causing some issues in extreme numerically unstable cases.

Fixes #86
