### 2023-02-21 #5158 @Nikoleta-v3
👋🏻 @ben-marshall and  @danpage would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "SCALib: A Side-Channel Analysis Library". You can find more information at the top of this Github issue 🆙
