### 2023-02-27 de0366a
Add clear_belief debug parameter
