### 2023-02-28 #5196 @editorialbot
[REVIEW]: SCALib: A Side-Channel Analysis Library
**Submitting author:** @cassiersg (<a href="http://orcid.org/0000-0001-5426-9345">Gaëtan Cassiers</a>)
**Repository:** https://github.com/simple-crypto/SCALib
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.5
