### 2023-02-28 1878246
SASCA arbitrary factor: initial implementation
