+++ src/scalib_ext/scalib/src/sasca/factor_graph.rs
@@ -28,6 +28,7 @@ new_id!(EdgeId, EdgeVec);
 pub type EdgeSlice<T> = index_vec::IndexSlice<EdgeId, [T]>;
 pub type PublicId = usize;
 pub type TableId = usize;
+pub type GenFactorId = usize;
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub(super) struct Var {
@@ -41,8 +42,6 @@ pub(super) struct Factor {
     pub(super) multi: bool,
     // res is first element (if there is a res), operands come next
     pub(super) edges: IndexMap<VarId, EdgeId>,
-    // Is the result a variable (and not a public) ?
-    pub(super) has_res: bool,
     // May not be allowed for all factor kinds
     pub(super) publics: Vec<(PublicId, bool)>,
 }
@@ -52,36 +51,59 @@ impl Factor {
         *self.edges.get_index(var_order).unwrap().0
     }
     pub(super) fn res_id(&self) -> Option<VarId> {
-        self.has_res.then_some(self.var_id(0))
+        if let FactorKind::Assign { has_res, .. } = &self.kind {
+            has_res.then_some(self.var_id(0))
+        } else {
+            None
+        }
     }
 }
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub(super) enum FactorKind<T = TableId> {
+pub(super) enum ExprFactor {
     AND { vars_neg: Vec<bool> },
     XOR,
     NOT,
     ADD,
     MUL,
-    LOOKUP { table: T },
+    LOOKUP { table: TableId },
 }
 
-impl FactorKind {
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub(super) enum GenFactorOperand {
+    Var(usize, bool),
+    Pub(usize),
+}
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub(super) enum FactorKind {
+    Assign {
+        expr: ExprFactor,
+        // Is the result a variable (and not a public) ?
+        has_res: bool,
+    },
+    GenFactor {
+        id: GenFactorId,
+        operands: Vec<GenFactorOperand>,
+    },
+}
+
+impl ExprFactor {
     fn merge(&self, a: ClassVal, b: ClassVal, nc: usize) -> ClassVal {
         match self {
-            FactorKind::AND { vars_neg: _ } => a & b,
-            FactorKind::XOR => a ^ b,
-            FactorKind::ADD => (((a as u64) + (b as u64)) % (nc as u64)) as ClassVal,
-            FactorKind::MUL => (((a as u64) + (b as u64)) % (nc as u64)) as ClassVal,
-            FactorKind::NOT | FactorKind::LOOKUP { .. } => unreachable!(),
+            Self::AND { vars_neg: _ } => a & b,
+            Self::XOR => a ^ b,
+            Self::ADD => (((a as u64) + (b as u64)) % (nc as u64)) as ClassVal,
+            Self::MUL => (((a as u64) + (b as u64)) % (nc as u64)) as ClassVal,
+            Self::NOT | Self::LOOKUP { .. } => unreachable!(),
         }
     }
     fn neutral(&self, nc: usize) -> ClassVal {
         match self {
-            FactorKind::AND { vars_neg: _ } => (nc - 1) as ClassVal,
-            FactorKind::XOR | FactorKind::ADD => 0,
-            FactorKind::MUL => 1,
-            FactorKind::NOT | FactorKind::LOOKUP { .. } => unreachable!(),
[…]
