+++ src/scalib_ext/scalib/src/sasca/belief_propagation.rs
@@ -2,14 +2,23 @@ use std::sync::Arc;
 
 use thiserror::Error;
 
+use super::factor_graph as fg;
 use super::factor_graph::{
-    EdgeId, EdgeSlice, EdgeVec, Factor, FactorId, FactorKind, FactorVec, Node, Table, VarId, VarVec,
+    EdgeId, EdgeSlice, EdgeVec, ExprFactor, Factor, FactorGraph, FactorId, FactorKind, FactorVec,
+    Node, PublicValue, Table, VarId, VarVec,
 };
-use super::{Distribution, FactorGraph, PublicValue};
+use super::Distribution;
+use ndarray::s;
 
 // TODO improvements
 // - use a pool for Distribution allocations (can be a simple Vec storing them), to avoid frequent
 // allocations
+
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub enum GenFactor {
+    Single(ndarray::ArrayD<f64>),
+    Multi(Vec<ndarray::ArrayD<f64>>),
+}
 
 // Workaround since the plans are not Serialize of Debug
 #[derive(Clone, serde::Serialize, serde::Deserialize)]
@@ -60,10 +69,12 @@ pub struct BPState {
     graph: std::sync::Arc<FactorGraph>,
     // number of parallel executions for PARA vars
     nmulti: u32,
-    // one public for every factor. Set to 0 if not relevant.
-    public_values: FactorVec<PublicValue>,
-    // public value for each factor
+    // list of public values
+    public_values: Vec<PublicValue>,
+    // one public value for every factor. Set to 0 if not relevant.
     pub_reduced: FactorVec<PublicValue>,
+    // generalized factors values
+    gen_factors: Vec<GenFactor>,
     // evidence for each var
     evidence: VarVec<Distribution>,
     // current proba for each var
@@ -96,6 +107,7 @@ impl BPState {
         graph: std::sync::Arc<FactorGraph>,
         nmulti: u32,
         public_values: Vec<PublicValue>,
+        gen_factors: Vec<GenFactor>,
     ) -> Self {
         let var_state: VarVec<_> = graph
             .vars
@@ -117,10 +129,11 @@ impl BPState {
             var_state,
             graph,
             nmulti,
-            public_values: FactorVec::from_vec(public_values),
+            public_values,
             pub_reduced,
             cyclic,
             plans,
+            gen_factors,
         }
     }
     pub fn is_cyclic(&self) -> bool {
@@ -231,15 +244,29 @@ impl BPState {
             };
         }
         match &factor.kind {
-            FactorKind::AND { .. } => {
-                prop_factor!(factor_gen_and, &self.pub_reduced[factor_id])
-            }
-            FactorKind::XOR => prop_factor!(factor_xor, &self.pub_reduced[factor_id]),
-            FactorKind::NOT => prop_factor!(factor_not, (self.graph.nc - 1) as u32),
-            FactorKind::ADD => prop_factor!(factor_add, &self.pub_reduced[factor_id], &self.plans),
-            FactorKind::MUL => prop_factor!(factor_mul, &self.pub_reduced[factor_id]),
-            FactorKind::LOOKUP { table } => {
-                prop_factor!(factor_lookup, &self.graph.tables[*table])
+            FactorKind::Assign { expr, .. } => match expr {
+                ExprFactor::AND { .. } => {
+                    prop_factor!(factor_gen_and, &self.pub_reduced[factor_id])
+                }
+                ExprFactor::XOR => prop_factor!(factor_xor, &self.pub_reduced[factor_id]),
+                ExprFactor::NOT => prop_factor!(factor_not, (self.graph.nc - 1) as u32),
[…]

[…]
