+++ src/scalib_ext/scalib/src/sasca/bp_compute.rs
@@ -232,11 +232,13 @@ impl Distribution {
         let num_arr = numerator
             .value()
             .expect("update_dampen needs full dist num");
-
+        let v: f64 = self_arr.iter().sum::<f64>();
         if let Some(denom_arr) = denominator.value() {
-            azip!((x in &mut self_arr, y in &num_arr, z in &denom_arr) *x = alpha * (y/(z + MIN_PROBA)) + (1.0 - alpha)* *x)
+            let u: f64 = num_arr.iter().sum::<f64>() / denom_arr.iter().sum::<f64>();
+            azip!((x in &mut self_arr, y in &num_arr, z in &denom_arr) *x = (alpha/u) * (y/(z + MIN_PROBA)) + ((1.0 - alpha)/v) * *x)
         } else {
-            azip!((x in &mut self_arr, y in &num_arr) *x = alpha * (y) + (1.0 - alpha)* *x)
+            let u: f64 = num_arr.iter().sum::<f64>();
+            azip!((x in &mut self_arr, y in &num_arr) *x = (alpha/u) * (y) + ((1.0 - alpha)/v) * *x)
         }
     }
 
