### 2023-03-01 9aff13a
Make ContextExecutor public, improve get_context() error msg.
