### 2023-03-01 6725576
Make ContextExecutor public, improve get_context() error msg.
