### 2023-03-03 06ec35f
Fix numerical underflow in BPState with multiple traces
The product of the likelihoods was taken without normalization. The
addition of constants did therefore not help, and actually led to wrong
(i.e. close to uniform) results.
