### 2023-03-03 f727be5
Fix numerical underflow in BPState with multiple traces
The product of the likelihoods was taken without normalization. The
addition of constants did therefore not help, and actually led to wrong
(i.e. close to uniform) results.
