### 2023-03-03 b724623
Move cargo test from tox to Makefile
