### 2023-03-03 6e8bc03
Move cargo test from tox to Makefile
