+++ tox.ini
@@ -13,7 +13,6 @@ setenv =
 deps =
     -r{toxinidir}/tests/requirements.txt
 commands =
-    cargo test --workspace --manifest-path {toxinidir}/src/scalib_ext/Cargo.toml
     pip freeze
     pytest {tty:--color=yes} {posargs}
 
@@ -26,7 +25,6 @@ deps = {[base]deps}
 commands = {[base]commands}
 package = wheel
 wheel_build_env = .pkg
-allowlist_externals = cargo
 
 [testenv:dev]
 description = Editable install with debug build.
@@ -62,6 +60,7 @@ skip_install = True
 changedir = .
 deps =
     black ~= 23.0
+allowlist_externals = cargo
 commands =
     black --version
     black {posargs} src/scalib tests examples
