+++ src/scalib/attacks/factor_graph.py
@@ -4,6 +4,7 @@ import numpy as np
 import numpy.typing as npt
 
 from scalib import _scalib_ext
+from scalib.config import get_config
 
 __all__ = ["FactorGraph", "BPState"]
 
@@ -233,8 +234,8 @@ class BPState:
             Whether to clear beliefs between vars -> factors. Setting to False can help debugging. Default value is True.
         """
         if initialize_states:
-            self._inner.propagate_all_vars(alpha, clear_beliefs)
-        self._inner.propagate_loopy_step(it, alpha, clear_beliefs)
+            self._inner.propagate_all_vars(get_config(), alpha, clear_beliefs)
+        self._inner.propagate_loopy_step(it, get_config(), alpha, clear_beliefs)
 
     def bp_acyclic(
         self,
@@ -255,7 +256,9 @@ class BPState:
         clear_evidence:
             Drop the evidence for the variables, once used in the algorithm.
         """
-        self._inner.propagate_acyclic(dest, clear_intermediates, clear_evidence)
+        self._inner.propagate_acyclic(
+            dest, clear_intermediates, clear_evidence, get_config()
+        )
 
     def get_distribution(self, var: str) -> Optional[npt.NDArray[np.float64]]:
         r"""Returns the current distribution of a variable `var`.
@@ -360,7 +363,7 @@ class BPState:
             Whether to clear beliefs between vars -> factors. Setting to False can help debugging. Default value is True.
 
         """
-        return self._inner.propagate_var(var, alpha, clear_beliefs)
+        return self._inner.propagate_var(var, get_config(), alpha, clear_beliefs)
 
     def propagate_factor(self, factor: str):
         """Run belief propagation on the given factor.
@@ -374,7 +377,7 @@ class BPState:
             Identifier of the variable.
 
         """
-        return self._inner.propagate_factor_all(factor)
+        return self._inner.propagate_factor_all(factor, get_config())
 
     def debug(self):
         """Debug-print the current state."""
