+++ src/scalib_ext/scalib-py/Cargo.toml
@@ -17,16 +17,16 @@ hellib = ["ranklib/hellib"]
 blis = ["scalib/blis"]
 
 [dependencies]
-rayon = "1.5.0"
-ndarray = { version = "0.15.1", features = ["rayon"] }
-numpy = "0.18.0"
-indicatif = "0.17.3"
-num_cpus = "1.15.0"
-bincode = "1.3.3" # Serialization for pickle support
+rayon = "1.7"
+ndarray = { version = "0.15", features = ["rayon"] }
+numpy = "0.18"
+indicatif = "0.17"
+num_cpus = "1.15"
+bincode = "1.3" # Serialization for pickle support
 
 [dependencies.pyo3]
-version = "0.18.0"
-features = ["extension-module"]
+version = "0.18"
+features = ["extension-module", "abi3-py38"]
 
 [dependencies.ranklib]
 path = "../ranklib"
