+++ src/scalib_ext/scalib-py/Cargo.toml
@@ -5,7 +5,7 @@ authors = [
     "Olivier Bronchain <olivier.bronchain@uclouvain.be>",
     "Gaëtan Cassiers <gaetan.cassiers@uclouvain.be>"
 ]
-edition = "2018"
+edition = "2021"
 
 [lib]
 name = "scalib_py"
