+++ src/scalib_ext/scalib/Cargo.toml
@@ -39,7 +39,7 @@ geigen = { path = "../geigen" }
 # Numerical libs
 ndarray = { version = "0.15", features = ["rayon", "approx", "serde"] }
 blis-src = { version = "0.2.1", default-features=false, features=["static", "pthreads"], optional = true }
-blis-sys2 = { git="https://github.com/cassiersg/blis-sys2", version = "0.2.1", optional = true }
+blis-sys2 = { version = "0.2.1", optional = true }
 matrixmultiply = "0.3"
 nalgebra = "0.32"
 nshare = { version="0.9", features=["nalgebra", "ndarray"] }
