### 2023-03-05 5b0d4d0
Add Mac Os to the CI and fix uArch checking bug
We don't compile Mac OS x86_64 with avx2 since github runners are too
old for that, so if we used it, we couldn't run tests.
