+++ src/scalib/__init__.py
@@ -15,8 +15,7 @@ from .version import version as __version__
 
 cpu = cpuinfo.get_cpu_info()
 
-
-if REQUIRE_AVX2 and cpu["arch"] == "X68_64" and "avx2" not in cpu["flags"]:
+if REQUIRE_AVX2 and cpu["arch"] == "X86_64" and "avx2" not in cpu["flags"]:
     raise ImportError(
         "SCALib has been compiled with AVX2 instructions, which are not "
         + "supported by your CPU or OS. See "
