### 2023-03-05 59db081
Add Mac Os to the CI and fix uArch checking bug
We don't compile Mac OS x86_64 with avx2 since github runners are too
old for that, so if we used it, we couldn't run tests.
