### 2023-03-05 7cd2d44
setup.cfg: upgrade deprecated field
