+++ setup.cfg
@@ -7,7 +7,7 @@ url = https://github.com/simple-crypto/scalib
 author= Olivier Bronchain, Gaëtan Cassiers
 author_email = olivier.bronchain@uclouvain.be, gaetan.cassiers@uclouvain.be
 license = GNU AGPL v3+
-license_file = COPYING
+license_files = COPYING
 
 [options]
 
