+++ tests/test_factorgraph.py
@@ -918,18 +918,24 @@ def test_dampening_nochange():
     distri_b = np.array([[0.8, 0.2]])
     graph = FactorGraph(graph)
     n = 1
-    bp_state_no_dampen = BPState(graph, n)
-    bp_state_no_dampen.set_evidence("a", distri_a)
-    bp_state_no_dampen.set_evidence("b", distri_b)
-    bp_state_full_dampen = BPState(graph, n)
-    bp_state_full_dampen.set_evidence("a", distri_a)
-    bp_state_full_dampen.set_evidence("b", distri_b)
+    bp_no_dampen = BPState(graph, n)
+    bp_no_dampen.set_evidence("a", distri_a)
+    bp_no_dampen.set_evidence("b", distri_b)
+    bp_dampen = BPState(graph, n)
+    bp_dampen.set_evidence("a", distri_a)
+    bp_dampen.set_evidence("b", distri_b)
 
-    bp_state_no_dampen.bp_loopy(5, initialize_states=False)
-    bp_state_full_dampen.bp_loopy(5, initialize_states=False, alpha=1.0)
+    for v in graph.vars():
+        bp_dampen.propagate_var(v)
+        bp_no_dampen.propagate_var(v)
+    prev = bp_no_dampen.get_belief_from_var("a", "s1")
+    alpha = 0.5
+    bp_dampen.bp_loopy(1, False, alpha=alpha, clear_beliefs=False)
+    bp_no_dampen.bp_loopy(1, False, clear_beliefs=False)
+    print(bp_no_dampen.get_belief_from_var("a", "s1"))
     assert np.allclose(
-        bp_state_no_dampen.get_distribution("x"),
-        bp_state_full_dampen.get_distribution("x"),
+        ((1 - alpha) * prev + (alpha) * bp_no_dampen.get_belief_from_var("a", "s1")),
+        bp_dampen.get_belief_from_var("a", "s1"),
     )
 
 
@@ -976,31 +982,49 @@ def test_dampening_correctness():
 
     fg = FactorGraph(factor_graph)
     alpha = 0.9
-    bp_dampen = BPState(fg, 1, public_values={"IV": 0xC})
-    bp_no_dampen = BPState(fg, 1, public_values={"IV": 0xC})
+    bp_dampen = BPState(fg, 2, public_values={"IV": 0xC})
+    bp_no_dampen = BPState(fg, 2, public_values={"IV": 0xC})
     for k in fg.vars():
-        d = make_distri(16, 1)
+        d = make_distri(16, 2)
         bp_no_dampen.set_evidence(k, distribution=d)
         bp_dampen.set_evidence(k, distribution=d)
 
     for v in fg.vars():
         bp_dampen.propagate_var(v)
         bp_no_dampen.propagate_var(v)
-    prev = bp_no_dampen.get_belief_from_var("A", "P7")
-    prev2 = bp_no_dampen.get_belief_from_var("C", "P4")
-    prev3 = bp_no_dampen.get_belief_from_var("K0", "P1")
+    for v in fg.factors():
+        bp_dampen.propagate_factor(v)
+        bp_no_dampen.propagate_factor(v)
+    prevs_vars = {}
+    prevs_factors = {}
+    for factor in fg._inner.factor_names():
+        for var in fg._inner.factor_scope(factor):
+            k = (var, factor)
+            prevs_vars[k] = normalize_distr(
+                bp_no_dampen.get_belief_from_var(var, factor)
+            )
+            prevs_factors[k] = normalize_distr(
+                bp_no_dampen.get_belief_to_var(var, factor)
+            )
     bp_dampen.bp_loopy(1, False, alpha=alpha, clear_beliefs=False)
     bp_no_dampen.bp_loopy(1, False, clear_beliefs=False)
 
-    assert np.allclose(
[…]
