### 2023-03-29 84abf84
Fix generic factor bug, add test
