+++ src/scalib/modeling/rldaclassifier.py
@@ -1,0 +1,197 @@
+import numpy as np
+import numpy.typing as npt
+
+from scalib import _scalib_ext
+from scalib.config import get_config
+
+
+class RLDAClassifier:
+    r"""Models the leakage :math:`\mathbf{l}` with :math:`n_s` dimensions using
+    the regression based linear discriminant analysis classifier (RLDA). [1]_
+
+    In a nutshell, this model performs LDA with the class means modelled as
+    linear regression based on the :math:`n_b` bits of the class value.
+    Compared to the :class:`scalib.modeling.LDAClassifier`, this model will
+    perform better when the number of classes is large and/or there are few
+    profiling traces.
+
+    Internally, it first estimates the coefficients of the regression then estimates
+    the projection matrix to reduce the dimensionality of the gaussian template to
+    :math:`p` dimensions. A second projection is applied such that the covariance
+    matrix is identity.
+
+    It is then able to predict the leakage likelihood
+
+    .. math::
+        \hat{\mathsf{f}}[\mathbf{l}|X=x] =
+        \alpha
+        \exp\left(
+        -\frac{1}{2} \lVert\mathbf{W}^T\mathbf{l} - \mathbf{A}\mathbf\beta(x)\rVert^2
+        \right).
+
+    Where :math:`\mathbf{W}` is the projection matrix, :math:`\mathbf{A}` the projected
+    regression coefficients, and :math:`\mathbf{\beta(x)}` the coefficients of :math:`x`.
+    The parameter :math:`\alpha = 1/\sqrt{(2\pi)^p\lvert\hat\Sigma_\mathbf{W}}\rvert` does
+    not need to be calculated as it will get canceled out when applying Bayes' law.
+
+    :class:`RLDAClassifier` provides the probability for each of the :math:`2^{n_b}`
+    classes with :meth:`predict_proba`.
+
+    >>> from scalib.modeling import RLDAClassifier
+    >>> import numpy as np
+    >>> traces_model = np.random.randint(0,256,(5000,10),dtype=np.int16)
+    >>> labels_model = np.random.randint(0,256,(5000,1),dtype=np.uint64)
+    >>> rlda = RLDAClassifier(8, 10, 1, 3)
+    >>> rlda.fit_u(traces_model, labels_model)
+    >>> rlda.solve()
+    >>> traces_test = np.random.randint(0,256,(5000,10),dtype=np.int16)
+    >>> prs = rlda.predict_proba(traces_test, 0)
+
+    Notes
+    -----
+    .. [1] "Efficient Regression-Based Linear Discriminant Analysis for Side-Channel Security Evaluations",
+        G. Cassiers, H. Devillez, F. Standaert, B. Udvarhelyi, TCHES 2023 - Issue 3.
+    """
+
+    def __init__(self, nb: int, ns: int, nv: int, p: int):
+        """
+        Parameters
+        ----------
+        nb:
+            Number of bits of the profiled variables.
+        ns:
+            Number of dimensions in the leakage (trace length).
+        nv:
+            Number of variables to profile
+        p:
+            Number of dimensions in the linear subspace.
+        """
+        self._ns = ns
+        self._nv = nv
+        self._inner = _scalib_ext.RLDA(nb, ns, nv, p)
+        self._solved = False
+
+    def fit_u(self, l: npt.NDArray[np.int16], x: npt.NDArray[np.uint64], gemm_mode=1):
+        """Update statistical model estimates with additional data.
+
+        This can be called multiple times, the state is accumulated.
+
+        Parameters
+        ----------
+        l : array_like, int16
+            Array that contains the traces. Shape `(n,ns)`.
+        x : array_like, uint64
+            Labels for each trace. Shape `(n,nv)`.
[…]
