+++ src/scalib_ext/scalib-py/src/lib.rs
@@ -8,8 +8,10 @@ use std::error::Error;
 
 mod belief_propagation;
 mod factor_graph;
+mod information;
 mod lda;
 mod ranking;
+mod rlda;
 mod snr;
 mod thread_pool;
 mod ttest;
@@ -118,6 +120,9 @@ fn _scalib_ext(py: Python, m: &PyModule) -> PyResult<()> {
     m.add_class::<ttest::MTtest>()?;
     m.add_class::<lda::LDA>()?;
     m.add_class::<lda::LdaAcc>()?;
+    m.add_class::<rlda::RLDA>()?;
+    m.add_class::<rlda::RLDAClusteredModel>()?;
+    m.add_class::<information::ItEstimator>()?;
     m.add_class::<factor_graph::FactorGraph>()?;
     m.add_class::<factor_graph::BPState>()?;
     m.add_class::<thread_pool::ThreadPool>()?;
