+++ src/scalib_ext/scalib/src/rlda.rs
@@ -1,0 +1,687 @@
+use crate::ScalibError;
+
+use geigen::Geigen;
+use kdtree::{distance::squared_euclidean, KdTree};
+use ndarray::{
+    s, Array1, Array2, Array3, Array4, ArrayView1, ArrayView2, ArrayViewMut1, ArrayViewMut2,
+    ArrayViewMut3, Axis, NewAxis, Zip,
+};
+use nshare::{ToNalgebra, ToNdarray1, ToNdarray2};
+use rayon::prelude::*;
+use serde::{Deserialize, Serialize};
+use std::cmp::min;
+use std::convert::TryInto;
+use std::ops::{AddAssign, SubAssign};
+
+// NBITS_CHUNKS defines the size in bits of the precomputed mean chunks. By default, 8bit chunks are used.
+const NBITS_CHUNK: usize = 8;
+const SIZE_CHUNK: usize = 1 << NBITS_CHUNK;
+
+/// RLDA contains accumulator, solver and probability predictor.
+///
+///
+/// Implementaion of RLDA
+///
+/// This module computes the projection matrices and projected regression coeficients of a Regression-based Linear Discriminant analysis.
+/// It first calculates a linear regression on nb+1 coefficients (+1 is fot the intercept), then during solve phase, computes regression coefficients in the high-dimensional space, the scatter matrices and the projection to a lower dimensionality matrix.
+/// It then applies a second projection to obtain an identity covariance matrix. It applies the 2 projections to the high-dimensional regression coefficients.
+///
+/// Finally, it precomputes parts of the 2**nb means in chunks, used in the prediction phase.
+///
+/// In more details,
+/// During fitting, each label is mapped to an (nb+1) coefficient array corresponding to the bits of the label + the intercept.
+/// A 0 bit is mapped to -1 and a 1 bit is mapped to 1.
+/// The intercept is always mapped to 1 and is at the beginning of the array.
+/// The matrix X can be defined as the collection of the n coefficient arrays used for profiling and has shape (n,nb+1).
+/// In practice it is never computed, instead, for each profiled variable and each trace :
+/// the accumulator xtx takes the outer product of the coefficient array with itself,
+/// xty takes the outer product of the coefficient array and the leakage,
+/// and scatter accumulates the outer product of each trace with itself.
+///
+/// This precomputation is done in order to avoid computation on all the coefficients for each class for a small increase in state size.
+/// The chunks are computed using NBITS_CHUNK coefficients at a time (except for the first one, which includes the intercept).
+///
+/// We have the mean for a class as follows (for each variable v and projected sample p) :
+/// mu[v,class,p] sum_i=0^nb+1 proj_coefs[v,p,i]*int2mul(class,ext_bit(i)
+///
+/// The sum is splitted into chunks as follows :
+/// mu_chunks[v,chunk_id,class,p] = sum_i=0^NBITS_CHUNK proj_coefs[v,NBITS_CHUNK*chunk_id+i,p]*int2mul(class,ext_bit(NBITS_CHUNK*chunk_id+i)
+/// The mean is then rewritten as follows :
+/// mu[v,class,p] = sum_i=0^n_chunks mu_chunks[v, i, (class>>(NBITS_CHUNK*i))&SIZE_CHUNK ,p]
+///
+/// During prediction phase, probability estimation is done for all the 2**nb classes based on the leakage traces.
+/// Prediction is done by generating for each class the squared distance between the mean of the class and the leakage.
+#[derive(Serialize, Deserialize)]
+pub struct RLDA {
+    /// Number of samples in trace
+    pub ns: usize,
+    /// Number of bits
+    pub nb: usize,
[…]
