### 2023-03-30 bbd144c
Add RLDAClassifier and RLDAInformationEstimator
Co-authored-by: Balazs Udvarhelyi <balazs.udvarhelyi@uclouvain.be>
