+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Not released
 v0.5.5 (2023/05/30)
 -------------------
 
+* Add ``scalib.modeling.RLDAClassifier`` and ``scalib.metrics.RLDAInformationEstimator``.
 * Fix ``Distribution`` size bug in belief propagation computation (#102).
 * Do not include debug symbols in release builds (makes wheel much smaller).
 * ``Ttest``: do not crash on non-contiguous/fortran-order arrays
@@ -154,7 +155,7 @@ v0.1.0 (2021/04/16)
 -------------------
 
 * Initial release, with the following features:
-  * LDA and Gaussian templates modelling
+  * LDA and Gaussian templates modeling
   * SNR
   * T-test any order (for TLVA)
   * Soft Analytical Side-Channel Attack (SASCA)
