### 2023-03-30 f97e55d
Fix generic factor bug
