+++ src/scalib_ext/scalib/src/sasca/belief_propagation.rs
@@ -798,13 +798,13 @@ fn factor_gen_factor<'a>(
                             todo!("Negated operands on generalized factors not yet implemented.");
                         }
                         let distr = &belief_from_var[factor.edges[*var_idx]];
-                        let mut new_gen_factor: ndarray::ArrayD<f64> = ndarray::ArrayD::zeros(gen_factor.slice_axis(ndarray::Axis(i), ndarray::Slice::new(0, Some(1), 1)).shape());
+                        let mut new_gen_factor: ndarray::ArrayD<f64> = ndarray::ArrayD::zeros(gen_factor.slice_axis(ndarray::Axis(op_idx), ndarray::Slice::new(0, Some(1), 1)).shape());
                         if let Some(distr) = distr.value() {
-                            for (d, gf) in distr.slice(s![i,..]).iter().zip(gen_factor.axis_iter(ndarray::Axis(i))) {
+                            for (d, gf) in distr.slice(s![i,..]).iter().zip(gen_factor.axis_chunks_iter(ndarray::Axis(op_idx), 1)) {
                                 new_gen_factor.scaled_add(*d, &gf);
                             }
                         } else {
-                            for gf in gen_factor.axis_iter(ndarray::Axis(i)) {
+                            for gf in gen_factor.axis_chunks_iter(ndarray::Axis(op_idx), 1) {
                                 new_gen_factor += &gf;
                             }
                         }
