### 2023-04-03 eb169f4
Make threadpool creation lazy.
This avoids bugs where the default threadpool is created, then the
process forks (e.g. MultiprocessPool), since forking and threads don't
mix in POSIX.
We now only create the rayon pool when it is first used.
This will not solve all bugs, but we cannot do much better (that's a
Python interpreter bug -- https://bugs.python.org/issue46464).
