### 2023-04-09 #5196 @nicolaimueller
Hi all, I am done with my review. The checklist can be found above. Additionally, I want to give some more detailed comments:

I recently had the opportunity to test SCALib on a Ubuntu 22.10 subsystem and I am delighted to report that I did not encounter any issues related to the use of a subsystem, which is not always the case. Overall, my experience with the library was very positive, and I found it to be both well-designed and well-documented.

The installation process was straightforward, and I successfully installed SCALib using pip without any issues. The readme also provided valuable information about potential pitfalls, such as outdated versions of pip, which I found helpful.

During the testing of the provided example, I encountered no problems. The library's textual feedback was clear and well-formatted, making it easy to understand what was happening at each step.

One aspect of SCALib that particularly intrigued me was the benchmark results. Although I understand that these results can be highly dependent on the setup, I was still curious to try to reproduce them. Unfortunately, I encountered an issue where the link to scabench was no longer working, and I was prompted for authentication when I attempted to clone it. It would be helpful if the authors could clarify how to reproduce the benchmark results or fix the missing link if possible.

Although I could not find a specific "statement of need" section, I did not find it to be a major issue, as I felt that all of the necessary information was still provided in the readme.

Overall, I believe that SCALib is a well-designed and well-documented library. The authors have taken care to address all of the JOSS requirements, and the documentation is both comprehensive and nicely formatted. Although I encountered a few minor issues during my testing, they did not detract from my overall positive impression of the library.

As this is my first review for JOSS, please let me know if there is anything missing or if a more detailed answer to any particular point is required.
