### 2023-04-11 3052a06
Fix underflow and bug in Factorkind
