+++ src/scalib_ext/scalib/src/sasca/bp_compute.rs
@@ -386,9 +386,9 @@ impl Distribution {
             let (sum, min) = d
                 .iter()
                 .fold((0.0f64, 0.0f64), |(sum, min), x| (sum + x, min.min(*x)));
-            let offset = -min + MIN_PROBA;
-            let norm_f = 1.0 / (sum + offset * d.len() as f64);
-            d.mapv_inplace(|x| (x + offset) * norm_f);
+            let offset = -min;
+            let norm_f = 1.0 / (dbg!((sum + offset) + MIN_PROBA) * d.len() as f64);
+            d.mapv_inplace(|x| ((x + offset) + MIN_PROBA) * norm_f);
         })
     }
     pub fn make_non_zero_signed(&mut self) {
