### 2023-04-15 #5196 @JannikZeitschner
Hi everyone! Once again please excuse the time it took for me to review! I am done now with my finished checklist above. I also would like to give some additional notes:

I tested SCALib in WSL2 with Ubuntu 20.04 running. The installation was extremely easy as the authors made a good job to point out possible pitfalls and ensuring that almost everything regarding installation is handled internally. The documentation is well written and easy to understand, while always providing compact examples for a quick understanding how to use certain functionality.

As Nico already pointed out there is no "statement of need" section in the documentation (at least I could not finde one), but I also agree that this is not a major issue due to the overall effort of providing very good explanations and descriptions even for non-specialists.

At first I encountered the same error as Nico when trying to execute the benchmarking results, but managed to "fix" the problem I think, which would probably also answer the question by @cassiersg:
The link https://github.com/cassiersg/SCABench works fine, but running `git clone https://github.com/simple-crypto/scabench` returns an error as it can not find the repository. However, changing the command to `git clone https://github.com/cassiersg/SCABench` worked for me, i.e., I think just the owner of the repository changed. After running `cd SCABench` followed by `make` I could execute the benchmarks without a problem. I would like to highlight the simplicity of executing the benchmarks. As it compares the functionality with two other libraries the benchmark takes care of potential dependencies and interaction with the other libraries.

Just a very minor remark: I did find two typos in the readme:
In
> SCALib should be useful for any side-channel practitionner

I think it should be "practitioner" and under the chapter "Versioning policy" I think it is "semantic versioning".

Overall I think SCALib is a very useful library that can help improve many tasks in the field of side-channel analysis. The authors address common challenges (among others size of processed data and speed-up thanks to parallelization), while providing an easy to use interface both regarding the documentation and practical use. In my opinion all JOSS requirements were taken into account.

As this is my first JOSS review as well, please let me know if anything is missing or you need further information.
