### 2023-04-25 #5196 @cassiersg
Answering to @nicolaimueller review:

- The link in the SCABench README is now fixed, following the instructions there should work. (https://github.com/cassiersg/SCABench#usage)
- Could you check the software paper generated by the bot [here](https://github.com/openjournals/joss-reviews/issues/5196#issuecomment-1447713725) ? There is a "statement of need" section.

Answering to @JannikZeitschner review:

- There is no explicit "statement of need" section in the documentation, but this content appears in the readme (https://github.com/simple-crypto/SCALib) and on the homepage of the documentation (https://scalib.readthedocs.io/en/stable/).
- The typos are fixed.
