### 2023-04-26 fcd301f
fix issue in ADD factor
