+++ tests/test_factorgraph.py
@@ -428,6 +428,50 @@ def test_ADD():
     assert np.allclose(distri_z_ref, distri_z)
 
 
+def test_ADD2():
+    """
+    Test ADD between distributions
+    """
+    nc = 16
+    n = 1
+    distri_x = np.zeros((n, nc))  # make_distri(nc, n)
+    distri_y = np.zeros((n, nc))  # make_distri(nc, n)
+
+    distri_x[:, 0] = 1.0
+    distri_y[:, 0] = 1.0
+
+    graph = f"""
+        # some comments
+        NC {nc}
+        PROPERTY F: z = x+y
+        VAR MULTI z
+        VAR MULTI x
+        VAR MULTI y
+
+        """
+
+    graph = FactorGraph(graph)
+    bp_state = BPState(graph, n)
+    bp_state.set_evidence("x", distri_x)
+    bp_state.set_evidence("y", distri_y)
+    print(bp_state.debug())
+    bp_state.bp_loopy(1, True)
+    print(bp_state.debug())
+    bp_state.bp_loopy(10, False)
+    print(bp_state.debug())
+    distri_z = bp_state.get_distribution("z")
+
+    distri_z_ref = np.zeros(distri_z.shape)
+    msg = np.zeros(distri_z.shape)
+
+    for x in range(nc):
+        for y in range(nc):
+            distri_z_ref[:, (x + y) % nc] += distri_x[:, x] * distri_y[:, y]
+
+    distri_z_ref = (distri_z_ref.T / np.sum(distri_z_ref, axis=1)).T
+    assert np.allclose(distri_z_ref, distri_z)
+
+
 def test_add_cst():
     """
     Test ADD of a constant
