+++ src/scalib_ext/scalib/src/sasca/belief_propagation.rs
@@ -618,7 +618,7 @@ fn factor_add<'a>(
         // We do take the product of all factors then divide because some factors could be zero.
         let mut dest_fft = Vec::with_capacity(dest.len());
         for (e, taken) in factor.edges.values().zip(taken_dest.iter()) {
-            let v = if *taken {
+            if *taken {
                 let mut fft_e = ndarray::Array2::zeros((nmulti, nc / 2 + 1));
                 belief_from_var[*e].fft_to(
                     fft_input_scratch.as_mut_slice(),
@@ -627,7 +627,6 @@ fn factor_add<'a>(
                     plans,
                 );
                 dest_fft.push(fft_e);
-                dest_fft.last().unwrap()
             } else {
                 belief_from_var[*e].fft_to(
                     fft_input_scratch.as_mut_slice(),
@@ -635,24 +634,30 @@ fn factor_add<'a>(
                     fft_scratch.as_mut_slice(),
                     plans,
                 );
-                &fft_tmp
-            };
-            if acc_fft_init {
-                acc_fft *= v;
-            } else {
-                acc_fft.assign(&v);
-                acc_fft_init = true;
+
+                if acc_fft_init {
+                    acc_fft *= &fft_tmp;
+                } else {
+                    acc_fft.assign(&fft_tmp);
+                    acc_fft_init = true;
+                }
             }
             if clear_incoming {
                 belief_from_var[*e].reset();
             }
         }
+
         // This could be done in O(l) instead of O(l^2) where l=dest.len() by
         // better caching product computations.
         let mut fft_scratch = plans.c2r.make_scratch_vec();
         return (0..dest.len())
             .map(move |i| {
-                let mut res = acc_fft.clone();
+                let mut res = if acc_fft_init {
+                    acc_fft.clone()
+                } else {
+                    ndarray::Array2::ones(acc_fft.raw_dim())
+                };
+
                 for (j, fft_op) in dest_fft.iter().enumerate() {
                     if j != i {
                         res *= fft_op;
