### 2023-04-26 4c60a0d
fix issue in ADD factor
