### 2023-04-27 b256379
Fix incorrect size of distribution in BP.
Matrix size computation was incorrect when a Distribution was Uniform.

Fixes #102
