### 2023-04-27 9a84d17
Fix incorrect size of distribution in BP.
Matrix size computation was incorrect when a Distribution was Uniform.

Fixes #102
