+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 Not released
 ------------
 
+* Fix ``Distribution`` size bug in belief propagation computation (#102).
+
 v0.5.4 (2023/04/26)
 -------------------
 
