+++ tox.ini
@@ -14,7 +14,7 @@ deps =
     -r{toxinidir}/tests/requirements.txt
 commands =
     pip freeze
-    pytest {tty:--color=yes} {posargs}
+    pytest {tty:--color=yes -k "test_ADD3"} {posargs}
 
 [testenv]
 description =
