### 2023-05-08 f685148
Draft sparse generic factor
