+++ src/scalib_ext/scalib/src/sasca/belief_propagation.rs
@@ -7,7 +7,7 @@ use super::factor_graph::{
     EdgeId, EdgeSlice, EdgeVec, ExprFactor, Factor, FactorGraph, FactorId, FactorKind, FactorVec,
     Node, PublicValue, Table, VarId, VarVec,
 };
-use super::Distribution;
+use super::{ClassVal,Distribution};
 use ndarray::s;
 
 // TODO improvements
@@ -16,8 +16,13 @@ use ndarray::s;
 
 #[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
 pub enum GenFactor {
-    Single(ndarray::ArrayD<f64>),
-    Multi(Vec<ndarray::ArrayD<f64>>),
+    Single(GenFactorInner),
+    Multi(Vec<GenFactorInner>),
+}
+#[derive(Debug, Clone, serde::Serialize, serde::Deserialize)]
+pub enum GenFactorInner {
+    Dense(ndarray::ArrayD<f64>),
+    SparseFunctional(ndarray::Array2<ClassVal>),
 }
 
 // Workaround since the plans are not Serialize of Debug
@@ -803,53 +808,103 @@ fn factor_gen_factor<'a>(
                 GenFactor::Single(x) => x,
                 GenFactor::Multi(x) => &x[i],
             };
-            assert_eq!(gen_factor.shape().len(), operands.len());
-            // First slice the array with the constants.
-            let gen_factor = gen_factor.slice_each_axis(|ax| match operands[ax.axis.index()] {
-                fg::GenFactorOperand::Var(_, _) => ndarray::Slice::new(0, None, 1),
-                fg::GenFactorOperand::Pub(pub_idx) => {
-                    let mut pub_val = public_values[factor.publics[pub_idx].0].get(i) as isize;
-                    if factor.publics[pub_idx].1 {
-                        if nc.is_power_of_two() {
-                            pub_val = !pub_val;
-                        } else {
-                            // TODO Check that we enforce this at graph creation time and return a proper error.
-                            panic!("Cannot negate operands with non-power-of-two number of classes.");
+            match gen_factor {
+                GenFactorInner::Dense(gen_factor) => {
+                    assert_eq!(gen_factor.shape().len(), operands.len());
+                    // First slice the array with the constants.
+                    let gen_factor = gen_factor.slice_each_axis(|ax| match operands[ax.axis.index()] {
+                        fg::GenFactorOperand::Var(_, _) => ndarray::Slice::new(0, None, 1),
+                        fg::GenFactorOperand::Pub(pub_idx) => {
+                            let mut pub_val = public_values[factor.publics[pub_idx].0].get(i) as isize;
+                            if factor.publics[pub_idx].1 {
+                                if nc.is_power_of_two() {
+                                    pub_val = !pub_val;
+                                } else {
+                                    // TODO Check that we enforce this at graph creation time and return a proper error.
+                                    panic!("Cannot negate operands with non-power-of-two number of classes.");
+                                }
+                            }
+                            ndarray::Slice::new(pub_val, Some(pub_val+1), 1)
+                        }
+                    });
+                    let mut gen_factor = gen_factor.to_owned();
+                    for (op_idx, op) in operands.iter().enumerate() {
+                        if op_idx != dest_idx {
+                            if let fg::GenFactorOperand::Var(var_idx, neg) = op {
+                                if *neg {
+                                    todo!("Negated operands on generalized factors not yet implemented.");
+                                }
+                                let distr = &belief_from_var[factor.edges[*var_idx]];
[…]
