+++ src/scalib_ext/scalib/src/sasca/mod.rs
@@ -9,7 +9,7 @@ pub use bp_compute::Distribution;
 pub type ClassVal = u32;
 type NamedList<T> = indexmap::IndexMap<String, T>;
 
-pub use belief_propagation::{BPError, BPState, GenFactor};
+pub use belief_propagation::{BPError, BPState, GenFactor, GenFactorInner};
 pub use factor_graph::{EdgeId, FactorGraph, FactorId, PublicValue, VarId};
 pub use fg_build::GraphBuildError;
 
