+++ src/scalib_ext/scalib-py/src/factor_graph.rs
@@ -149,6 +149,23 @@ fn pyobj2pubs<'a>(
     }
 }
 
+fn pyobj2genfactor_inner(py: Python, obj: &PyAny) -> PyResult<sasca::GenFactorInner> {
+    let kind: u32 = obj.getattr("kind")?.extract()?;
+    let dense: u32 = obj.getattr("GenericFactorKind")?.getattr("DENSE")?.extract()?;
+    let sparse_functional: u32 = obj.getattr("GenericFactorKind")?.getattr("SPARSE_FUNCTIONNAL")?.extract()?;
+    if kind == dense {
+        let factor: &numpy::PyArrayDyn<f64> = obj.getattr("factor")?.extract()?;
+        let factor = factor.readonly().as_array().as_standard_layout().into_owned();
+        Ok(sasca::GenFactorInner::Dense(factor))
+    } else if kind == sparse_functional {
+        let factor: &numpy::PyArray2<sasca::ClassVal> = obj.getattr("factor")?.extract()?;
+        let factor = factor.readonly().as_array().as_standard_layout().into_owned();
+        Ok(sasca::GenFactorInner::SparseFunctional(factor))
+    } else {
+        Err(PyValueError::new_err(("Unknown kind", obj.getattr("kind")?.to_object(py))))
+    }
+}
+
 fn pyobj2factors<'a>(
     py: Python,
     gen_factors: PyObject,
@@ -168,17 +185,15 @@ fn pyobj2factors<'a>(
                         name
                     )));
                 }
-                let obj: Vec<&numpy::PyArrayDyn<f64>> = gf.extract(py)?;
+                let obj: Vec<&PyAny> = gf.extract(py)?;
                 Ok(sasca::GenFactor::Multi(
                     obj.into_iter()
-                        .map(|obj| obj.readonly().as_array().as_standard_layout().into_owned())
-                        .collect(),
+                        .map(|obj| pyobj2genfactor_inner(py, obj))
+                        .collect::<Result<Vec<_>,_>>()?,
                 ))
             } else {
-                let obj: &numpy::PyArrayDyn<f64> = gf.extract(py)?;
-                Ok(sasca::GenFactor::Single(
-                    obj.readonly().as_array().as_standard_layout().into_owned(),
-                ))
+                let obj: &PyAny = gf.extract(py)?;
+                Ok(sasca::GenFactor::Single(pyobj2genfactor_inner(py, obj)?))
             }
         })
         .collect::<Result<Vec<sasca::GenFactor>, PyErr>>()?;
