+++ src/scalib_ext/scalib-py/src/factor_graph.rs
@@ -151,18 +151,32 @@ fn pyobj2pubs<'a>(
 
 fn pyobj2genfactor_inner(py: Python, obj: &PyAny) -> PyResult<sasca::GenFactorInner> {
     let kind: u32 = obj.getattr("kind")?.extract()?;
-    let dense: u32 = obj.getattr("GenericFactorKind")?.getattr("DENSE")?.extract()?;
-    let sparse_functional: u32 = obj.getattr("GenericFactorKind")?.getattr("SPARSE_FUNCTIONNAL")?.extract()?;
+    let dense: u32 = obj.getattr("GenFactorKind")?.getattr("DENSE")?.extract()?;
+    let sparse_functional: u32 = obj
+        .getattr("GenFactorKind")?
+        .getattr("SPARSE_FUNCTIONAL")?
+        .extract()?;
     if kind == dense {
         let factor: &numpy::PyArrayDyn<f64> = obj.getattr("factor")?.extract()?;
-        let factor = factor.readonly().as_array().as_standard_layout().into_owned();
+        let factor = factor
+            .readonly()
+            .as_array()
+            .as_standard_layout()
+            .into_owned();
         Ok(sasca::GenFactorInner::Dense(factor))
     } else if kind == sparse_functional {
         let factor: &numpy::PyArray2<sasca::ClassVal> = obj.getattr("factor")?.extract()?;
-        let factor = factor.readonly().as_array().as_standard_layout().into_owned();
+        let factor = factor
+            .readonly()
+            .as_array()
+            .as_standard_layout()
+            .into_owned();
         Ok(sasca::GenFactorInner::SparseFunctional(factor))
     } else {
-        Err(PyValueError::new_err(("Unknown kind", obj.getattr("kind")?.to_object(py))))
+        Err(PyValueError::new_err((
+            "Unknown kind",
+            obj.getattr("kind")?.to_object(py),
+        )))
     }
 }
 
@@ -189,7 +203,7 @@ fn pyobj2factors<'a>(
                 Ok(sasca::GenFactor::Multi(
                     obj.into_iter()
                         .map(|obj| pyobj2genfactor_inner(py, obj))
-                        .collect::<Result<Vec<_>,_>>()?,
+                        .collect::<Result<Vec<_>, _>>()?,
                 ))
             } else {
                 let obj: &PyAny = gf.extract(py)?;
