### 2023-05-09 1128cab
Do not include debug symbols in release builds (makes wheel much smaller)
