### 2023-05-09 597db4c
Ttest: do not crash on non-contiguous arrays
