+++ tests/test_ttest.py
@@ -112,3 +112,20 @@ def test_ttest_d6():
     ttest.fit_u(traces, labels)
     t = ttest.get_ttest()
     assert np.allclose(t_ref, t, rtol=1e-3)
+
+
+def test_ttest_data_transpose():
+    ns = 100
+    d = 1
+    nc = 2
+    n = 200
+
+    m = np.random.randint(0, 2, (nc, ns))
+    traces = np.asfortranarray(np.random.randint(0, 10, (n, ns), dtype=np.int16))
+    labels = np.random.randint(0, nc, n, dtype=np.uint16)
+    traces += m[labels]
+
+    t_ref = reference(traces, labels, d)
+    ttest = Ttest(ns, 1)
+    with pytest.raises(ValueError):
+        ttest.fit_u(traces, labels)
