### 2023-05-10 84f00d5
normalize before multiply
