+++ src/scalib_ext/scalib/src/sasca/belief_propagation.rs
@@ -1,5 +1,5 @@
 use std::sync::Arc;
-
+use std::slice;
 use thiserror::Error;
 
 use super::factor_graph as fg;
@@ -250,7 +250,8 @@ impl BPState {
             ($f:ident, $($arg:expr),*) => {
                 {
                     let it = $f(factor, &mut self.belief_from_var, dest, clear_incoming, $($arg,)*);
-                    for (distr, dest) in it.zip(dest.iter()) {
+                    for (mut distr, dest) in it.zip(dest.iter()) {
+                        distr.normalize();
                         self.belief_to_var[factor.edges[dest]] = distr;
                     }
                 }
@@ -803,111 +804,192 @@ fn factor_gen_factor<'a>(
         let dest_idx = factor.edges.get_index_of(dest).unwrap();
         let mut distr = belief_from_var[factor.edges[dest_idx]].clone();
         distr.ensure_full();
-        for i in 0..nmulti {
-            let gen_factor = match gen_factor {
-                GenFactor::Single(x) => x,
-                GenFactor::Multi(x) => &x[i],
-            };
-            match gen_factor {
-                GenFactorInner::Dense(gen_factor) => {
-                    assert_eq!(gen_factor.shape().len(), operands.len());
-                    // First slice the array with the constants.
-                    let gen_factor = gen_factor.slice_each_axis(|ax| match operands[ax.axis.index()] {
-                        fg::GenFactorOperand::Var(_, _) => ndarray::Slice::new(0, None, 1),
-                        fg::GenFactorOperand::Pub(pub_idx) => {
-                            let mut pub_val = public_values[factor.publics[pub_idx].0].get(i) as isize;
-                            if factor.publics[pub_idx].1 {
-                                if nc.is_power_of_two() {
-                                    pub_val = !pub_val;
-                                } else {
-                                    // TODO Check that we enforce this at graph creation time and return a proper error.
-                                    panic!("Cannot negate operands with non-power-of-two number of classes.");
-                                }
-                            }
-                            ndarray::Slice::new(pub_val, Some(pub_val+1), 1)
-                        }
-                    });
-                    let mut gen_factor = gen_factor.to_owned();
-                    for (op_idx, op) in operands.iter().enumerate() {
-                        if op_idx != dest_idx {
-                            if let fg::GenFactorOperand::Var(var_idx, neg) = op {
-                                if *neg {
-                                    todo!("Negated operands on generalized factors not yet implemented.");
-                                }
-                                let distr = &mut belief_from_var[factor.edges[*var_idx]];
-                                distr.normalize();
-                                let mut new_gen_factor: ndarray::ArrayD<f64> = ndarray::ArrayD::zeros(gen_factor.slice_axis(ndarray::Axis(op_idx), ndarray::Slice::new(0, Some(1), 1)).shape());
+        dbg!(&distr);
+        
+        match gen_factor {
+            GenFactor::Single(GenFactorInner::SparseFunctional(gen_factor_inner)) => {
+               // let gen_factor_inner = gen_factor_inner.to_owned();
+                for op_values in gen_factor_inner.outer_iter() {
+                    let mut res = ndarray::Array1::ones((nmulti,));
+                    for (op_idx, (op, val)) in operands.iter().zip(op_values.iter()).enumerate() {
+                        match op {
+                            fg::GenFactorOperand::Var(var_idx, neg) => {
+                                let distr = &belief_from_var[factor.edges[*var_idx]];
+                                // For uniform, we implicitly multiply by 1.0
                                 if let Some(distr) = distr.value() {
-                                    for (d, gf) in distr.slice(s![i,..]).iter().zip(gen_factor.axis_chunks_iter(ndarray::Axis(op_idx), 1)) {
-                                        new_gen_factor.scaled_add(*d, &gf);
[…]
