### 2023-05-17 #5196 @Nikoleta-v3
@cassiersg  thank you for your patience. The topic is outside my field of expertise, so my comments (mostly) focus on the technical aspects.

Indeed the package was installed without an issue - I managed to run your script and your test suite (I can also see that you have a CI setup) 👍

One issue that I have is regarding the documentation. In your `examples` folder you have one example that uses the Signal-to-Noise Ratio (SNR)  metric, the LDAClassifier and the function `SASCAGraph` for the attacks which is deprecated why are you not using `FactorGraph` instead?

My issue is that there is only one example that is showing some of the functionalities of the library. For all the methods, metrics etc implemented within the package you have very nice docstrings and examples in the online documentation but they feel a bit hidden.

I would recommend to include some more examples either in the examples file or in the documentation itself, or even simpler in the README (in the `examples` folder) you can let users know about the other functionalities and point them to correct sections in the documentation.

The testing information also felt a bit hidden but I don’t have a recommendation regarding this.

Regarding the paper there are two typos:

- Line 42: `an` evaluation
- Line 69: relies `on` an efficient

One other comment is regarding your reviewers. I asked you to recommend some reviewers, and I made the final call about who was assigned to your submission. However, as I can see here: https://github.com/ChairImpSec/PROLEAD  you have collaborated with both of your reviewers. This is something you should have mentioned as a potential conflict of interest.
