+++ src/scalib_ext/scalib/src/sasca/belief_propagation.rs
@@ -1,5 +1,5 @@
-use std::sync::Arc;
 use std::slice;
+use std::sync::Arc;
 use thiserror::Error;
 
 use super::factor_graph as fg;
@@ -800,41 +800,46 @@ fn factor_gen_factor<'a>(
     nc: usize,
 ) -> impl Iterator<Item = Distribution> + 'a {
     let fg::FactorKind::GenFactor { operands, .. } = &factor.kind else { unreachable!() };
-    let res: Vec<Distribution> = dest.iter().map(|dest| {
-        let dest_idx = factor.edges.get_index_of(dest).unwrap();
-        let mut distr = belief_from_var[factor.edges[dest_idx]].clone();
-        distr.ensure_full();
-        dbg!(&distr);
-        
-        match gen_factor {
-            GenFactor::Single(GenFactorInner::SparseFunctional(gen_factor_inner)) => {
-               // let gen_factor_inner = gen_factor_inner.to_owned();
-                for op_values in gen_factor_inner.outer_iter() {
-                    let mut res = ndarray::Array1::ones((nmulti,));
-                    for (op_idx, (op, val)) in operands.iter().zip(op_values.iter()).enumerate() {
-                        match op {
-                            fg::GenFactorOperand::Var(var_idx, neg) => {
-                                let distr = &belief_from_var[factor.edges[*var_idx]];
-                                // For uniform, we implicitly multiply by 1.0
-                                if let Some(distr) = distr.value() {
-                                    res *= &distr.slice(s![..,*val as usize]);
+    let res: Vec<Distribution> = dest
+        .iter()
+        .map(|dest| {
+            let dest_idx = factor.edges.get_index_of(dest).unwrap();
+            let mut distr = belief_from_var[factor.edges[dest_idx]].clone();
+            distr.ensure_full();
+            let mut tdest = distr.value_mut().unwrap();
+            tdest.fill(0.0);
+
+            match gen_factor {
+                GenFactor::Single(GenFactorInner::SparseFunctional(gen_factor_inner)) => {
+                    // let gen_factor_inner = gen_factor_inner.to_owned();
+                    for op_values in gen_factor_inner.outer_iter() {
+                        let mut res = ndarray::Array1::ones((nmulti,));
+                        for (op_idx, (op, val)) in operands.iter().zip(op_values.iter()).enumerate()
+                        {
+                            match op {
+                                fg::GenFactorOperand::Var(var_idx, neg) => {
+                                    let distr = &belief_from_var[factor.edges[*var_idx]];
+                                    // For uniform, we implicitly multiply by 1.0
+                                    if let Some(distr) = distr.value() {
+                                        res *= &distr.slice(s![.., *val as usize]);
+                                    }
                                 }
-                            }
-                            fg::GenFactorOperand::Pub(pub_idx) => {
-                                let mut pub_val = &public_values[factor.publics[*pub_idx].0];
-                                if factor.publics[*pub_idx].1 {
-                                    todo!();
-                                } else {
-                                    match pub_val {
-                                        PublicValue::Single(class_val) => {
-                                            if *class_val != *val {
-                                                res.fill(0.0);
+                                fg::GenFactorOperand::Pub(pub_idx) => {
+                                    let mut pub_val = &public_values[factor.publics[*pub_idx].0];
+                                    if factor.publics[*pub_idx].1 {
+                                        todo!();
+                                    } else {
+                                        match pub_val {
+                                            PublicValue::Single(class_val) => {
+                                                if *class_val != *val {
+                                                    res.fill(0.0);
+                                                }
                                             }
-                                        }
-                                        PublicValue::Multi(class_val) => {
-                                            for i in 0..nmulti {
-                                                if class_val[i] != *val {
[…]
