### 2023-05-17 f29b530
Example: move to FactorGraph and add T-test.
