### 2023-05-17 7c87709
Example: move to FactorGraph and add T-test.
