+++ examples/aes_simulation.py
@@ -1,6 +1,6 @@
-from scalib.metrics import SNR
+from scalib.metrics import SNR, Ttest
 from scalib.modeling import LDAClassifier
-from scalib.attacks import SASCAGraph
+from scalib.attacks import FactorGraph, BPState
 from scalib.postprocessing import rank_accuracy
 
 from utils import sbox, gen_traces
@@ -12,19 +12,36 @@ def main():
     npoi = 2
 
     # Parameters
-    std = 1
-    ntraces_a = 1000
+    std = 2
+    ntraces_a = 40
     ntraces_p = 20000
 
-    print("1. Generate simulated traces with parameters:")
+    print("1. Generate simulated traces (Hamming weight + Gaussian noise) with parameters:")
     print(f"    ntraces for profiling: {ntraces_p}")
     print(f"    ntraces for attack:    {ntraces_a}")
     print(f"    noise std:             {std}")
-    traces_p, labels_p = gen_traces(ntraces_p, std, random_key=1)
-    traces_a, labels_a = gen_traces(ntraces_a, std, random_key=0)
+    traces_p, labels_p = gen_traces(ntraces_p, std, random_key=True, random_plaintext=True)
+    traces_a, labels_a = gen_traces(ntraces_a, std, random_key=False, random_plaintext=True)
 
     _, ns = traces_p.shape
-    print("2. Profiling")
+
+    print("2. TVLA (not further used -- first leakage expected, but no second order leakage)")
+    # Ttest needs the lenght of the traces, and the test order (here, two)
+    ttest = Ttest(traces_p.shape[1], 2)
+    # Fix-vs-random TVLA (fixed-key vs random-key).
+    # We generate new datasets, as the previous ones use random plaintext, and we needed fixed plaintext.
+    traces_rk, _ = gen_traces(ntraces_p, std, random_key=True, random_plaintext=False)
+    traces_fk, _ = gen_traces(ntraces_p, std, random_key=False, random_plaintext=False)
+    # Ttest can be updated multiple times, and the order of the updates does not matter.
+    ttest.fit_u(traces_rk, np.zeros((traces_rk.shape[0],), dtype=np.uint16))
+    ttest.fit_u(traces_fk, np.ones((traces_fk.shape[0],), dtype=np.uint16))
+    tt_result = ttest.get_ttest()
+    print("")
+    print("    First-order t-statistic: ", *(f"{t:.02f}" for t in tt_result[0,:]))
+    print("    Second-order t-statistic:", *(f"{t:.02f}" for t in tt_result[1,:]))
+    print("")
+
+    print("3. Profiling")
     # For each of the 16 variables at the output of the Sbox (xi)
     # we build a model that store in the dictonnary models.
     #
@@ -37,7 +54,7 @@ def main():
     for i in range(16):
         models[f"x{i}"] = {}
 
-    print("    2.1 Compute snr for 16 Sbox outputs xi")
+    print("    3.1 Compute SNR for 16 Sbox outputs xi")
     # y array with xi values
     x = np.zeros((ntraces_p, 16), dtype=np.uint16)
     for i in range(16):
@@ -48,12 +65,12 @@ def main():
     snr.fit_u(traces_p, x)
     snr_val = snr.get_snr()
 
-    print(f"    2.2 Keep {npoi} POIs for each xi")
+    print("    3.2 Keep {npoi} POIs for each xi")
     # POI are the npoi indexes with the largest SNR
     for i in range(16):
         models[f"x{i}"]["poi"] = np.argsort(snr_val[i])[-npoi:]
 
-    print("    2.3 Build LDAClassifier for each xi")
+    print("    3.3 Build LDAClassifier for each xi")
     for i in range(16):
[…]
