### 2023-05-17 a26dd9c
paper: typos and affiliations/funding
