+++ paper/paper.md
@@ -14,14 +14,12 @@ authors:
     orcid: 0000-0001-5426-9345
     equal-contrib: true
     corresponding: true
-    affiliation: "1, 2, 3"
+    affiliation: "1, 2"
 affiliations:
  - name: UCLouvain, Belgium
    index: 1
  - name: Graz University of Technology, Austria
    index: 2
- - name: Lamarr Security Research, Austria
-   index: 3
 date: 16 February 2023
 bibliography: paper.bib
 
@@ -45,13 +43,13 @@ known behavior [@DBLP:conf/eurocrypt/StandaertMY09].
 
 For the statistical processing, we distinguish between two classes of attacks,
 based on the use of a profiling dataset.
-Such a dataset consists in leakage measurements on a device running the
+Such a dataset consists of leakage measurements on a device running the
 cryptographic algorithm with the known key.
 Profiled attacks use this data to fit a statistical model (or train a
 machine-learning model) of the device, while non-profiled attacks have to rely
 on *a priori* models and are therefore less powerful [@DBLP:conf/ches/ChariRR02].
 
-There are two main approaches to evaluating security of devices against side-channel attacks.
+There are two main approaches for evaluating the security of devices against side-channel attacks.
 First, attack-based evaluations try to attack the device and report their success or failure.
 In case of success, the main figure of merit is the number of traces (i.e.,
 number of executions of a cryptographic algorithm for which the leakage is
@@ -65,12 +63,12 @@ analyzing the effectiveness of a newly proposed countermeasure or analyzing a
 widely deployed device.
 In `SCALib`, we implement algorithms for commonly used metrics and methods in
 side-channel security evaluations, attack-based and evaluation-based.
-We however focus on the requirements of evaluations, and do not implement
+We focus on the requirements of evaluations and do not implement
 complete attacks when they are not needed to evaluate the security of a device.
 
 `SCALib` is distributed as a Python package and uses 16-bit integer `numpy` [@numpy] arrays
 for leakage traces.
-For the sake of efficiency, most algorithms are however implemented in Rust,
+For the sake of efficiency, most algorithms are implemented in Rust,
 allowing fine control of the memory accesses and enabling efficient
 parallelization.
 
@@ -81,36 +79,36 @@ Many of the algorithms used in side-channel security evaluations are well-known
 statistical techniques.
 For instance, the widely used TVLA methodology is based on the Welch t-test for
 the difference of means [@DBLP:conf/ches/SchneiderM15].
-Also when modeling the leakage, techniques such as Linear Discriminant Analysis
+Also, when modeling the leakage, techniques such as Linear Discriminant Analysis
 (LDA) [@DBLP:conf/ches/StandaertA08] can be used.
 While implementations of these algorithms are fairly easy to find, our use-case
 has a few particularities that motivate dedicated implementations.
-For example, the number of traces used in a evaluation can be very large,
+For example, the number of traces used in an evaluation can be very large,
 amounting to terabytes of data, hence incremental single-pass algorithms (that
 avoid the need to store and/or load multiple times the dataset) are highly
 desirable.
-Moreover, while the leakage samples are acquired at relatively low-resolution
+Moreover, while the leakage samples are acquired at a fairly low-resolution
 (8-bit to 16-bit integers), detection of very small effect sizes is
 needed, as they can potentially be exploited to mount an attack.
 Besides this requirement, leakage traces contain many points (typically
-thousands) and many metrics have to be computed for each of these points,
+thousands), and many metrics have to be computed for each of these points,
 providing parallelization opportunities.
 As a result of these characteristics, dedicated implementations can achieve
[…]
