### 2023-05-17 73fbd70
Example: show more of SCALib
