### 2023-05-17 0127190
Example: show more of SCALib
