+++ examples/aes_attack.py
@@ -16,30 +16,35 @@ def main():
     ntraces_a = 40
     ntraces_p = 20000
 
-    print("1. Generate simulated traces (Hamming weight + Gaussian noise) with parameters:")
+    print(
+        "1. Generate simulated traces (Hamming weight + Gaussian noise) with parameters:"
+    )
     print(f"    ntraces for profiling: {ntraces_p}")
     print(f"    ntraces for attack:    {ntraces_a}")
     print(f"    noise std:             {std}")
-    traces_p, labels_p = gen_traces(ntraces_p, std, random_key=True, random_plaintext=True)
-    traces_a, labels_a = gen_traces(ntraces_a, std, random_key=False, random_plaintext=True)
+    traces_p, labels_p = gen_traces(
+        ntraces_p, std, random_key=True, random_plaintext=True
+    )
+    traces_a, labels_a = gen_traces(
+        ntraces_a, std, random_key=False, random_plaintext=True
+    )
 
     _, ns = traces_p.shape
 
-    print("2. TVLA (not further used -- first leakage expected, but no second order leakage)")
-    # Ttest needs the lenght of the traces, and the test order (here, two)
-    ttest = Ttest(traces_p.shape[1], 2)
-    # Fix-vs-random TVLA (fixed-key vs random-key).
-    # We generate new datasets, as the previous ones use random plaintext, and we needed fixed plaintext.
-    traces_rk, _ = gen_traces(ntraces_p, std, random_key=True, random_plaintext=False)
-    traces_fk, _ = gen_traces(ntraces_p, std, random_key=False, random_plaintext=False)
-    # Ttest can be updated multiple times, and the order of the updates does not matter.
-    ttest.fit_u(traces_rk, np.zeros((traces_rk.shape[0],), dtype=np.uint16))
-    ttest.fit_u(traces_fk, np.ones((traces_fk.shape[0],), dtype=np.uint16))
-    tt_result = ttest.get_ttest()
-    print("")
-    print("    First-order t-statistic: ", *(f"{t:.02f}" for t in tt_result[0,:]))
-    print("    Second-order t-statistic:", *(f"{t:.02f}" for t in tt_result[1,:]))
-    print("")
+    print("2. POI selection with SNR")
+    print("    2.1 Compute SNR for 16 Sbox outputs xi")
+    # y array with xi values
+    x = np.zeros((ntraces_p, 16), dtype=np.uint16)
+    for i in range(16):
+        x[:, i] = labels_p[f"x{i}"]
+
+    # estimate SNR
+    snr = SNR(nc=nc, ns=ns, np=16)
+    snr.fit_u(traces_p, x)
+    snr_val = snr.get_snr()
+
+    print("    2.2 Select POIs with highest SNR.")
+    pois = [np.argsort(snr_val[i])[-npoi:] for i in range(16)]
 
     print("3. Profiling")
     # For each of the 16 variables at the output of the Sbox (xi)
@@ -50,33 +55,16 @@ def main():
     # 2. Select the Point-of-Interest based on the SNR
     # 3. Fit an LDA model for each of the xi
 
-    models = {}
+    print("    3.1 Build LDAClassifier for each xi")
+    models = []
     for i in range(16):
-        models[f"x{i}"] = {}
-
-    print("    3.1 Compute SNR for 16 Sbox outputs xi")
-    # y array with xi values
-    x = np.zeros((ntraces_p, 16), dtype=np.uint16)
-    for i in range(16):
-        x[:, i] = labels_p[f"x{i}"]
-
-    # estimate SNR
-    snr = SNR(nc=256, ns=ns, np=16)
-    snr.fit_u(traces_p, x)
-    snr_val = snr.get_snr()
[…]
