+++ examples/utils.py
@@ -33,7 +33,7 @@ def gen_traces(ntraces, std, random_key, random_plaintext):
         p = np.random.randint(0, 256, (ntraces, 16), dtype=np.uint8)
     else:
         p = np.random.randint(0, 256, (16,), dtype=np.uint8)
-        p = np.vstack(ntraces*[p])
+        p = np.vstack(ntraces * [p])
 
     # generate sensible data.
     # Input and output of the Sbox
