### 2023-05-19 048cf36
Remove intended DeprecationWarning's in tests
