+++ README.rst
@@ -16,6 +16,7 @@ The Side-Channel Analysis Library (SCALib) is a Python library that
 contains state-of-the-art tools for side-channel security evaluation.
 
 - **Documentation**: https://scalib.readthedocs.io/
+- **Examples**: `examples/ <examples/>`_
 - **Chat**: `https://matrix.to/#/#scalib:matrix.org`
 - **Source code**: https://github.com/simple-crypto/SCALib
 - **Bug reports/feature requests**: https://github.com/simple-crypto/SCALib/issues/new/choose
