### 2023-05-19 #5196 @cassiersg
@Nikoleta-v3  Thanks for the comments!

> One issue that I have is regarding the documentation. In your `examples` folder you have one example that uses the Signal-to-Noise Ratio (SNR) metric, the LDAClassifier and the function `SASCAGraph` for the attacks which is deprecated why are you not using `FactorGraph` instead?
>
> My issue is that there is only one example that is showing some of the functionalities of the library. For all the methods, metrics etc implemented within the package you have very nice docstrings and examples in the online documentation but they feel a bit hidden.
>
> I would recommend to include some more examples either in the examples file or in the documentation itself, or even simpler in the README (in the `examples` folder) you can let users know about the other functionalities and point them to correct sections in the documentation.
>
> The testing information also felt a bit hidden but I don’t have a recommendation regarding this.

Regarding the I indeed almost forgot it lately. I made the following changes:
- Use `FactorGraph` and `BPState` instead of `SASCAGraph`.
- Add two other example scripts.
- Improve `examples/README.rst`.

*All the changes discussed in this message are in https://github.com/simple-crypto/SCALib/pull/109 (I keep it open for now, in case further changes are needed).*

I tried to make the `exmples/` a good resource for beginner users, covering most of the features of the library, except for some advanced/niche ones, which are mentioned in `examples/README.rst` but not explained further.

I also tried to emphasize the links to the documentation to make it less "hidden", both in the top-level README and in `examples/README.rst`. A few useful links are also highlighted at the top of the top-level README. This hopefully helps making the testing information more accessible, but I don't want to emphasize it too much, as the test suite is mostly useful for developers (and CI).

>
> Regarding the paper there are two typos:
>
>     * Line 42: `an` evaluation
>
>     * Line 69: relies `on` an efficient

Thanks, fixed.

>
>
> One other comment is regarding your reviewers. I asked you to recommend some reviewers, and I made the final call about who was assigned to your submission. However, as I can see here: https://github.com/ChairImpSec/PROLEAD you have collaborated with both of your reviewers. This is something you should have mentioned as a potential conflict of interest.

I did not realize that this pull request could indeed be considered as a conflict of interest, I apologize for not mentioning it.

*Full COI disclosure:* The reviewers published https://eprint.iacr.org/2023/034, in which they found an issue in the artifact of one of my papers. Following that, I tried to reproduce their results and we exchanged five emails (I needed support to be able to run the tool) last januaray/february. As I had made a few trivial fixes to get the tool to build, I sent this pull request: https://github.com/ChairImpSec/PROLEAD/pull/1. These have been my only collaborations with the reviewers.
