### 2023-05-23 ebfae3d
Build dependencies versions and build commands.
