### 2023-05-30 #5196 @Nikoleta-v3
Thank you for your work @cassiersg and for clarifying the situation with the reviewers.

The `examples` folder has improved a lot 👍🏻 I just left some minor comments on the `pr`: https://github.com/simple-crypto/SCALib/pull/109#pullrequestreview-1446134438.
