### 2023-05-30 b58a67b
Examples: fix typos and outdated comments
