### 2023-05-30 #5196 @cassiersg
Thanks again @Nikoleta-v3 for reviewing the changes.

- I made a new release with the reviewed changes, with fixes according to latest feedback: [`v0.5.5`](https://github.com/simple-crypto/SCALib/releases/tag/v0.5.5).
- This release has been archived in Zenodo: https://zenodo.org/record/7985686 . DOI: 10.5281/zenodo.7985686
- I updated the paper metadata and acknowledgements (https://github.com/simple-crypto/SCALib/commit/ff78dca0a2b74b2936f55153ca574c7752f4f850).
