+++ paper/paper.bib
@@ -44,15 +44,15 @@
   editor    = {Palash Sarkar and
                Tetsu Iwata},
   title     = {Soft Analytical Side-Channel Attacks},
-  booktitle = {Advances in Cryptology - {ASIACRYPT} 2014 - 20th International Conference
-               on the Theory and Application of Cryptology and Information Security,
-               Kaoshiung, Taiwan, R.O.C., December 7-11, 2014. Proceedings, Part
-               {I}},
+  booktitle = {{Advances in Cryptology - Proceedings of 20th International Conference
+               on the Theory and Application of Cryptology and Information Security (ASIACRYPT} 2014), Part I}},
   series    = {Lecture Notes in Computer Science},
   volume    = {8873},
   pages     = {282--296},
   publisher = {Springer},
   year      = {2014},
+  month = dec,
+  location = {Kaoshiung, Taiwan, R.O.C.},
   url       = {https://doi.org/10.1007/978-3-662-45611-8_15},
   doi       = {10.1007/978-3-662-45611-8_15},
   timestamp = {Thu, 14 Oct 2021 10:07:58 +0200},
@@ -67,13 +67,14 @@
                Pankaj Rohatgi},
   title     = {Using Subspace-Based Template Attacks to Compare and Combine Power
                and Electromagnetic Information Leakages},
-  booktitle = {Cryptographic Hardware and Embedded Systems - {CHES} 2008, 10th International
-               Workshop, Washington, D.C., USA, August 10-13, 2008. Proceedings},
+  booktitle = {{Proceedings of 10th International Workshop on Cryptographic Hardware and Embedded Systems (CHES 2008)}},
   series    = {Lecture Notes in Computer Science},
   volume    = {5154},
   pages     = {411--425},
   publisher = {Springer},
   year      = {2008},
+  month = aug,
+  location = {Washington, D.C., USA},
   url       = {https://doi.org/10.1007/978-3-540-85053-3_26},
   doi       = {10.1007/978-3-540-85053-3_26},
   timestamp = {Thu, 14 Oct 2021 10:28:51 +0200},
@@ -88,13 +89,14 @@
                Helena Handschuh},
   title     = {Leakage Assessment Methodology - {A} Clear Roadmap for Side-Channel
                Evaluations},
-  booktitle = {Cryptographic Hardware and Embedded Systems - {CHES} 2015 - 17th International
-               Workshop, Saint-Malo, France, September 13-16, 2015, Proceedings},
+  booktitle = {{Proceedings of 17th International Workshop on Cryptographic Hardware and Embedded Systems (CHES 2015)}},
   series    = {Lecture Notes in Computer Science},
   volume    = {9293},
   pages     = {495--513},
   publisher = {Springer},
   year      = {2015},
+  month = sep,
+  location = {Saint-Malo, France},
   url       = {https://doi.org/10.1007/978-3-662-48324-4_25},
   doi       = {10.1007/978-3-662-48324-4_25},
   timestamp = {Thu, 31 Oct 2019 17:50:49 +0100},
@@ -110,12 +112,13 @@
                Axel Y. Poschmann},
   title     = {Simple Key Enumeration (and Rank Estimation) Using Histograms: An
                Integrated Approach},
-  booktitle = {Cryptographic Hardware and Embedded Systems - {CHES} 2016 - 18th International
-               Conference, Santa Barbara, CA, USA, August 17-19, 2016, Proceedings},
+  booktitle = {{Proceedings of 18th International Conference on Cryptographic Hardware and Embedded Systems - (CHES 2016)}},
   series    = {Lecture Notes in Computer Science},
   volume    = {9813},
   pages     = {61--81},
[…]
