+++ paper/paper.bib
@@ -157,7 +157,7 @@
   editor    = {Antoine Joux},
   title     = {A Unified Framework for the Analysis of Side-Channel Key Recovery
                Attacks},
-  booktitle = {Advances in Cryptology - Proceedings of 28th Annual International Conference on the Theory and Applications of Cryptographic Techniques (EUROCRYPT 200:9)}},
+  booktitle = {{Advances in Cryptology - Proceedings of 28th Annual International Conference on the Theory and Applications of Cryptographic Techniques (EUROCRYPT 200:9)}},
   series    = {Lecture Notes in Computer Science},
   volume    = {5479},
   pages     = {443--461},
